+++ extensions/hide-files/src/utils/constants.ts
@@ -1,0 +1,25 @@
+export enum LocalStorageKey {
+  LOCAL_HIDE_DIRECTORY = "localHideDirectory",
+  LOCAL_HIDE_TOGGLE = "localHideToggle",
+}
+
+export const imgExt = [
+  ".cr2",
+  ".cr3",
+  ".gif",
+  ".gif",
+  ".heic",
+  ".heif",
+  ".icns",
+  ".icon",
+  ".icons",
+  ".jpeg",
+  ".jpg",
+  ".jpg",
+  ".png",
+  ".raf",
+  ".raw",
+  ".svg",
+  ".tiff",
+  ".webp",
+];