+++ extensions/text-shortcuts/src/run-shortcut-now.tsx
@@ -1,14 +1,28 @@
 import React, { useState } from "react";
-import { Action, ActionPanel, Clipboard, Form, Icon, showToast, Toast } from "@raycast/api";
+import {
+  Action,
+  ActionPanel,
+  Clipboard,
+  Form,
+  getPreferenceValues,
+  Icon,
+  showHUD,
+  showToast,
+  Toast,
+} from "@raycast/api";
 import { handleLiveTemplate, runShortcut, Taction } from "./util/shortcut";
 import { fetchItemInput } from "./util/input";
 import { TactionActions, tactionForms } from "./create-shortcut";
+import { Preference } from "./util/utils";
 
 export default function CreateShortcut() {
   const [tactions, setTactions] = useState<Taction[]>([]);
+  const { closeMainWindow } = getPreferenceValues<Preference>();
 
   return (
-    <Form actions={<RunShortcutActions tactions={tactions} setTactions={setTactions} />}>
+    <Form
+      actions={<RunShortcutActions tactions={tactions} setTactions={setTactions} closeMainWindow={closeMainWindow} />}
+    >
       {tactionForms(tactions, setTactions)}
       <Form.Description text={"  ⌘D       ⌘E       ⌘N        ⌘R            ⌘T              ⌘L"} />
       <Form.Description text={"Delete | Coder | Case | Replace | Transform | Template"} />
@@ -19,9 +33,9 @@ export default function CreateShortcut() {
 function RunShortcutActions(props: {
   tactions: Taction[];
   setTactions: React.Dispatch<React.SetStateAction<Taction[]>>;
+  closeMainWindow: boolean;
 }) {
-  const tactions = props.tactions;
-  const setTactions = props.setTactions;
+  const { tactions, setTactions, closeMainWindow } = props;
   return (
     <ActionPanel>
       <Action
@@ -31,7 +45,11 @@ function RunShortcutActions(props: {
           const _inputItem = await fetchItemInput();
           const _runShortcut = runShortcut(_inputItem.content, handleLiveTemplate(tactions));
           await Clipboard.paste(_runShortcut);
-          await showToast(Toast.Style.Success, "Pasted result to active app!");
+          if (closeMainWindow) {
+            await showHUD("Pasted result to active app");
+          } else {
+            await showToast(Toast.Style.Success, "Pasted result to active app!");
+          }
         }}
       />