+++ extensions/music/CHANGELOG.md
@@ -1,8 +1,15 @@
 # Apple Music Changelog
+
+## [Fix] - 2022-05-15
+
+Fixed issue #1692
+
 ## [Track Rating / Fixes] - 2022-05-05
+
 - Fixed issue #1294
 - Added song rating #1568
 - Refactored some `AppleScript` code
+
 ## [Search/Play Playlist Feature] - 2022-03-17
 
 Added new search and play playlist feature to extension